requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
Brlapi==0.6.6
certifi==2018.11.29
cffi==1.12.0
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
command-not-found==0.3
constantly==15.1.0
contentful==1.11.4
contentful-management==2.7.0
cryptography==2.5
cssselect==1.0.3
cupshelpers==1.0
defer==1.0.6
distro-info==0.18
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
httplib2==0.9.2
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
krakenio==0.0.2
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.3.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
parsel==1.5.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pyasn1==0.1.9
pyasn1-modules==0.2.4
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
PyDispatcher==2.0.5
pygobject==3.26.1
PyHamcrest==1.9.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==19.0.0
pyRFC3339==1.0
python-apt==1.6.3+ubuntu1
python-dateutil==2.8.0
python-debian==0.1.32
python-dotenv==0.10.1
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
queuelib==1.5.0
reportlab==3.4.0
requests==2.21.0
requests-unixsocket==0.1.5
Scrapy==1.6.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.13.2
six==1.10.0
system-service==0.3
systemd-python==234
tinify==1.5.1
Twisted==18.9.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.3
visitor==0.1.3
w3lib==1.20.0
wadllib==1.3.2
Werkzeug==0.14.1
xkit==0.0.0
zope.interface==4.6.0
